+++ TDEngine2/include/graphics/CBaseGraphicsPipeline.h
@@ -10,6 +10,7 @@
 #include "../utils/Utils.h"
 #include "../utils/Types.h"
 #include "../core/CBaseObject.h"
+#include "../core/Serialization.h"
 #include "IGraphicsPipeline.h"
 
 
@@ -42,16 +43,24 @@ namespace TDEngine2
 			TDE2_API E_RESULT_CODE Init(IGraphicsContext* pGraphicsContext, const TGraphicsPipelineConfigDesc& pipelineConfig) override;
 
 			TDE2_API E_RESULT_CODE Bind() override;
+
+			TDE2_API const TGraphicsPipelineConfigDesc& GetConfig() const override;
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CBaseGraphicsPipeline)
 		protected:
-			IGraphicsContext*           mpGraphicsContext;
-			IGraphicsObjectManager*     mpGraphicsObjectManager;
+			IGraphicsContext*           mpGraphicsContext = nullptr;
+			IGraphicsObjectManager*     mpGraphicsObjectManager = nullptr;
 			TGraphicsPipelineConfigDesc mConfig;
 
 			TBlendStateId               mBlendStateHandle = TBlendStateId::Invalid;
 			TDepthStencilStateId        mDepthStencilStateHandle = TDepthStencilStateId::Invalid;
 			TRasterizerStateId          mRasterizerStateHandle = TRasterizerStateId::Invalid;
-
 	};
+
+
+	// \note The reason why there are defined Serialize/Deserialize for TGraphicsPipelineConfigDesc instead of making CBaseGraphicsPipeline ISerializable
+	// is that we should know config before we create a new instance of the type
+
+	template <> TDE2_API E_RESULT_CODE Serialize<TGraphicsPipelineConfigDesc>(class IArchiveWriter* pWriter, TGraphicsPipelineConfigDesc value);
+	template <> TDE2_API TResult<TGraphicsPipelineConfigDesc> Deserialize<TGraphicsPipelineConfigDesc>(IArchiveReader* pReader);
 }